requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart>=0.0.7,<0.0.10

openai-whisper==20231117
numpy==1.26.4
pydantic>=2.7,<3
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
